feat:处理健壮性问题

+++ user_del.sh
@@ -17,6 +17,12 @@ if ! id "$username" &>/dev/null; then
     exit 1
 fi
 
+# 检查用户是否有进程在运行
+if pgrep -u "${username}" &>/dev/null; then
+    echo "{\"success\": false, \"message\": \"该用户下存在桌面未关闭\"}"
+    exit 1
+fi
+
 # 强制终止用户进程
 pkill -u "${username}" &>/dev/null
